* fixing tests and code after overhauling the scan readers
