+++ mzLib/Readers/IDynamicDataConnection.cs
@@ -7,7 +7,6 @@ namespace Readers
 {
     public interface IDynamicDataConnection
     {
-        public string FilePath { get; }
         public MsDataScan GetOneBasedScanFromDynamicConnection(int oneBasedScanNumber, IFilteringParams filterParams = null);
         public void CloseDynamicConnection();
         public void InitiateDynamicConnection();
